Addressing pr review comments
